apps/api/pyproject.toml
[project]
name = "vaktram-api"
version = "0.1.0"
description = "Vaktram meeting intelligence API"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn[standard]>=0.30.0,<1.0.0",
    "pydantic>=2.9.0,<3.0.0",
    "pydantic-settings>=2.5.0,<3.0.0",
    "sqlalchemy[asyncio]>=2.0.35,<3.0.0",
    "asyncpg>=0.30.0,<1.0.0",
    "alembic>=1.14.0,<2.0.0",
    "greenlet>=3.0.0",
    "PyJWT>=2.9.0,<3.0.0",
    "supabase>=2.10.0,<3.0.0",
    "upstash-redis>=1.1.0,<2.0.0",
    "openai>=1.40.0,<2.0.0",
    "anthropic>=0.34.0,<1.0.0",
    "google-genai>=1.0.0",
    "groq>=0.9.0,<1.0.0",
    "boto3>=1.35.0,<2.0.0",
    "httpx>=0.27.0,<1.0.0",
    "cryptography>=43.0.0,<44.0.0",
    "python-multipart>=0.0.12",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.7.0",
    "mypy>=1.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
plugins = ["pydantic.mypy"]
ignore_missing_imports = true
